requirements.txt
Flask==3.0.3
feedparser==6.0.11
python-dotenv==1.0.1
SQLAlchemy==2.0.36
APScheduler==3.10.4
Jinja2==3.1.4
requests==2.32.3
PyYAML==6.0.2
bleach==6.2.0
python-dateutil==2.9.0
